Use existing add/removeEventListener for elements
This makes it possible to use can-define with HTML elements so that both
property events and DOM events can be listened to and fire correctly.

This is done by using the can-event/lifecycle/lifecycle mixin, and
passing it our existing addEventListener (where it exists), rather than
using canEvent.addEventListener as lifecycle did before.

Closes #145